Update height for photos + font specification
Update made to force the dimensions to stay 224x224 even when a photo
is not square. Hit this error when working on the Congressional Briefs
pages in Shorthand.

Another update made add but comment out the font-family specification.
This was needed in the Congressional Briefs pages because we were not
using the default font. Left in as a comment/flag to future users to
consider the font they're using and how to change it if they need to.